Move CueAvoid to pooltool.objects.cue
