+++ pooltool/ani/modes/aim.py
@@ -5,182 +5,9 @@ import pooltool.utils as utils
 import pooltool.ani.utils as autils
 
 from pooltool.ani.modes import Mode, action
-
-from panda3d.core import *
+from pooltool.objects.cue import CueAvoid
 
 import numpy as np
-
-
-class CueAvoid(object):
-    def __init__(self):
-        """Calculates minimum elevation required by cue stick to avoid colliding with balls and cushions
-
-        This class uses Panda3D collision detection to determine when the cue stick is intersecting
-        with a ball or cushion. Rather than use the built in collision solving (e.g.
-        https://docs.panda3d.org/1.10/python/reference/panda3d.core.CollisionHandlerPusher), which
-        tended to push the cue off of objects in arbitrary ways (such that the cue no longer pointed
-        at the cueing ball), I instead rely on geometry to solve the minimum angle that the cue
-        stick must be raised in order to avoid all collisions. At each step in AimMode.aim_task, if
-        the cue elevation is less than this angle, the elevation is automatically set to this
-        minimum.
-
-        Notes
-        =====
-        - This class has nothing to do with collisions that occurr during the shot evolution, e.g.
-          ball-ball collisions, ball-cushion collisions, etc. All of those are handled in events.py
-        """
-
-        self.min_theta = 0
-
-        # Declare frequently used nodes
-        self.avoid_nodes = {
-            'scene': render.find('scene'),
-            'cue_collision_node': self.cue.get_node('cue_cseg'),
-            'cue_stick_model': self.cue.get_node('cue_stick_model'),
-            'cue_stick': self.cue.get_node('cue_stick'),
-            'cue_stick_focus': self.cue.get_node('cue_stick_focus'),
-        }
-
-
-    def collision_task(self, task):
-        max_min_theta = 0
-
-        # Lay cue collision segment flat
-        self.avoid_nodes['cue_collision_node'].setR(-self.avoid_nodes['cue_stick_focus'].getR())
-
-        for entry in self.collision_handler.entries:
-            min_theta = self.process_collision(entry)
-            if min_theta > max_min_theta:
-                max_min_theta = min_theta
-
-        self.min_theta = max_min_theta
-        return task.cont
-
-
-    def process_collision(self, entry):
-        if not entry.has_surface_point():
-            # Not a collision we care about
-            return 0
-        elif entry.into_node.name.startswith('cushion'):
-            return self.process_cushion_collision(entry)
-        elif entry.into_node.name.startswith('ball'):
-            return self.process_ball_collision(entry)
-        else:
-            raise NotImplementedError(f"CueAvoid :: no collision solver for node {entry.into_node.name}")
-
-
-    def process_cushion_collision(self, entry):
-        cushion = self.get_cushion(entry)
-        cushion_height = cushion.p1[2]
-
-        # Point where cue center contacts collision plane
-        Px, Py, Pz = entry.getSurfacePoint(self.avoid_nodes['scene'])
-
-        # The tip of the cue stick
-        Ex, Ey, Ez = self.avoid_nodes['cue_stick_model'].getPos(self.avoid_nodes['scene'])
-
-        # Center ofthe cueing ball
-        Bx, By, Bz = self.avoid_nodes['cue_stick_focus'].getPos(self.avoid_nodes['scene'])
-
-        # The desired point where cue contacts collision plane, excluding cue width
-        Dx, Dy, Dz = Px, Py, cushion_height
-
-        # Center of aim
-        v = np.array([Ex-Px, Ey-Py, Ez-Pz])
-        u = utils.unit_vector(v)*self.avoid_nodes['cue_stick_model'].getX()
-        Fx, Fy, Fz = Ex + u[0], Ey + u[1], Ez + u[2]
-
-        min_theta = np.arctan2(Dz-Fz, np.sqrt((Dx-Fx)**2 + (Dy-Fy)**2))
-
-        # Correct for cue's cylindrical radius at collision point
-        # distance from cue tip (E) to desired collision point (D)
-        l = np.sqrt((Dx-Ex)**2 + (Dy-Ey)**2 + (Dz-Ez)**2)
-        cue_radius = self.get_cue_radius(l)
-        min_theta += np.arctan2(cue_radius, l)
-
-        return max(0, min_theta) * 180/np.pi
-
-
-    def process_ball_collision(self, entry):
-        min_theta = 0
-        ball = self.get_ball(entry)
-
-        if ball == self.cueing_ball:
-            return 0
-
-        scene = render.find('scene')
-
-        # Radius of transect
-        n = np.array(entry.get_surface_normal(render.find('scene')))
-        phi = ((self.avoid_nodes['cue_stick_focus'].getH() + 180) % 360) * np.pi/180
-        c = np.array([np.cos(phi), np.sin(phi), 0])
-        gamma = np.arccos(np.dot(n, c))
-        AB = (ball.R + self.cue.tip_radius)*np.cos(gamma)
-
-        # Center of blocking ball transect
-        Ax, Ay, _ = entry.getSurfacePoint(scene)
-        Ax -= (AB + self.cue.tip_radius)*np.cos(phi)
-        Ay -= (AB + self.cue.tip_radius)*np.sin(phi)
-        Az = ball.R
-
-        # Center of aim, leveled to ball height
-        Cx, Cy, Cz = self.avoid_nodes['cue_stick_focus'].getPos(scene)
-        axR = -self.avoid_nodes['cue_stick'].getY()
-        Cx += -axR*np.sin(phi)
-        Cy += axR*np.cos(phi)
-
-        AC = np.sqrt((Ax-Cx)**2 + (Ay-Cy)**2 + (Az-Cz)**2)
-        BC = np.sqrt(AC**2 - AB**2)
-        min_theta_no_english = np.arcsin(AB/AC)
-
-        # Cue tip point, no top/bottom english
-        m = self.avoid_nodes['cue_stick_model'].getX()
-        u = utils.unit_vector(np.array([-np.cos(phi), -np.sin(phi), np.sin(min_theta_no_english)]))
-        Ex, Ey, Ez = Cx + m*u[0], Cy + m*u[1], Cz + m*u[2]
-
-        # Point where cue contacts blocking ball, no top/bottom english
-        Bx, By, Bz = Cx + BC*u[0], Cy + BC*u[1], Cz + BC*u[2]
-
-        # Extra angle due to top/bottom english
-        BE = np.sqrt((Bx-Ex)**2 + (By-Ey)**2 + (Bz-Ez)**2)
-        bxR = self.avoid_nodes['cue_stick'].getZ()
-        beta = -np.arctan2(bxR, BE)
-        if beta < 0:
-            beta += 10*np.pi/180*(np.exp(bxR/BE)**2 - 1)
-
-        min_theta = min_theta_no_english + beta
-        return max(0, min_theta) * 180/np.pi
-
-
-    def get_cue_radius(self, l):
-        """Returns radius of cue at collision point, given collision point is distance l from cue tip"""
-
-        bounds = self.cue.get_node('cue_stick').get_tight_bounds()
-        L = bounds[1][0] - bounds[0][0] # cue length
-
-        r = self.cue.tip_radius
-        R = self.cue.butt_radius
-
-        m = (R - r)/L # rise/run
-        b = r # intercept
-
-        return m*l + b
-
-
-    def get_cushion(self, entry):
-        expected_suffix = 'cushion_cplane_'
-        into_node_path_name = entry.get_into_node_path().name
-        assert into_node_path_name.startswith(expected_suffix)
-        cushion_id = into_node_path_name[len(expected_suffix):]
-        return self.table.cushion_segments['linear'][cushion_id]
-
-
-    def get_ball(self, entry):
-        expected_suffix = 'ball_csphere_'
-        into_node_path_name = entry.get_into_node_path().name
-        assert into_node_path_name.startswith(expected_suffix)
-        ball_id = into_node_path_name[len(expected_suffix):]
-        return self.balls[ball_id]
 
 
 class AimMode(Mode, CueAvoid):
